modified:   deployers/universal_scripts/0_common_conf.sh 	modified:   deployers/universal_scripts/0_create_docker_machine.sh 	new file:   deployers/universal_scripts/1_4_initial_worker_2.sh 	new file:   deployers/universal_scripts/1_5_initial_worker_3.sh 	new file:   deployers/universal_scripts/1_6_initial_worker_4.sh 	new file:   deployers/universal_scripts/1_7_initial_worker_5.sh 	new file:   deployers/universal_scripts/1_8_create_overlay_network.sh 	modified:   deployers/universal_scripts/2_1_copy_pems.sh 	modified:   deployers/universal_scripts/2_2_pull_images.sh 	modified:   deployers/universal_scripts/2_3_copy_confs_and_set_sysctl_to_machines.sh 	new file:   deployers/universal_scripts/2_4_clear_all_volumes.sh 	new file:   deployers/universal_scripts/3_1_create_service_containers.sh 	new file:   deployers/universal_scripts/3_2_import_grafana_settings.sh 	modified:   deployers/universal_scripts/3_5_create_mongodb_services.sh 	modified:   deployers/universal_scripts/3_6_initiate_mongodb.sh 	modified:   deployers/universal_scripts/3_7_create_consumer_containers.sh 	modified:   deployers/universal_scripts/3_8_create_provider_containers.sh 	new file:   deployers/universal_scripts/3_9_create_analyze_containers.sh 	modified:   deployers/universal_scripts/4_1_stop_containers.sh 	modified:   deployers/universal_scripts/4_3_start_containers.sh 	modified:   deployers/universal_scripts/5_remove_containers.sh 	modified:   deployers/universal_scripts/6_remove_cluster.sh

+++ deployers/universal_scripts/5_remove_containers.sh
@@ -3,19 +3,30 @@ source ./0_common_conf.sh
 eval $(docker-machine env --swarm $Master_Host)
 docker rm -f $(docker ps -f name=act_sim -aq)
 docker rm -f $(docker ps -f name=rec_sim -aq)
-docker rm -f analyze
+docker rm -f clustering
+docker rm -f $(docker ps -f name=analyze -aq)
+docker rm -f $(docker ps -f name=model -aq)
 docker rm -f $(docker ps -f name=remove -aq)
 docker rm -f $(docker ps -f name=count -aq)
-docker rm -f reporter
-docker rm -f $(docker ps -f name=act_proc -aq)
-docker rm -f $(docker ps -f name=rec_proc -aq)
+docker rm -f portal
+
+#docker rm -f $(docker ps -f name=act_proc -aq)
+#docker rm -f $(docker ps -f name=rec_proc -aq)
 
 docker rm -f $(docker ps -f name=act_store -aq)
 docker rm -f $(docker ps -f name=rec_store -aq)
 
-docker rm -f $(docker ps -f name=logstash -aq)
-docker rm -f kibana
-docker rm -f $(docker ps -f name=elasticsearch -aq)
+docker rm -f $(docker ps -f name=cadvisor -aq)
+docker rm -f $(docker ps -f name=exporter -aq)
+docker rm -f redis_exporter
+
+if [ $with_elk ] ; then
+	echo "Remove ELK services.."
+	docker rm -f $(docker ps -f name=logstash -aq)
+	docker rm -f kibana
+	docker rm -f $(docker ps -f name=elasticsearch -aq)
+
+fi
 
 docker rm -f grafana
 docker rm -f prometheus
